Uncommented permission check in plugin checks

+++ events/message.js
@@ -28,7 +28,7 @@ module.exports = async (client, message) => {
 
 
   const prefix = client.guildData.has(message.guild.id, "prefix") ? client.guildData.get(message.guild.id, "prefix") : client.config.defaultSettings.prefix
-  if (await client.resolveUser(message.content.split(" ")[0]) === client.user) message.channel.send(
+  if (await client.resolveUser(message.content.split(" ")[0].id) === client.user.id) message.channel.send(
     new Discord.MessageEmbed()
     .setDescription("👋 **Hello " + message.author.toString() + ", my prefix is `" + prefix + '`. \nUse the command `help` for all of my commands!**')
     .setAuthor(message.author.tag, message.author.displayAvatarURL())
@@ -38,7 +38,7 @@ module.exports = async (client, message) => {
 
   if (client.plugins.get(message.guild.id, "invites")) {
     let inviteRegex = /(https?:\/\/)?(www\.)?(disc(ord)?\.(gg|li|me|io)|discordapp\.com\/invite|invite\.gg)\/.+/gi
-    if (inviteRegex.test(message.content) /*&& !message.member.hasPermission("MANAGE_GUILD")*/ ) {
+    if (inviteRegex.test(message.content) && !message.member.permissions.has("MANAGE_GUILD") ) {
       message.delete().catch(() => {})
       return message.channel.send(`${message.author.toString()}, ${client.emoji.misc.xmark} **You aren't allowed to send invites in this server**`).then(m => {
         setTimeout(() => {
@@ -48,7 +48,7 @@ module.exports = async (client, message) => {
     }
   } else if (client.plugins.get(message.guild.id, "links")) {
     let urlReg = /^(?:(?:https?|ftp):\/\/)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/\S*)?$/ig
-    if (urlReg.test(message.content) /* && !message.member.hasPermission("MANAGE_GUILD")*/ ) {
+    if (urlReg.test(message.content) && !message.member.permissions.has("MANAGE_GUILD") ) {
       message.delete().catch(() => {})
       return message.channel.send(`${message.author.toString()}, ${client.emoji.misc.xmark} **You aren't allowed to send links in this server**`).then(m => {
         setTimeout(() => {
@@ -59,7 +59,7 @@ module.exports = async (client, message) => {
   } else if (client.plugins.get(message.guild.id, "spoilers")) {
     let spoilerRegex = /\|\|.*?\|\|/gmi
     const match = message.content.match(spoilerRegex)
-    if (Array.isArray(match) && match.length >= 3) {
+    if (Array.isArray(match) && match.length >= 3 && !message.member.permissions.has("MANAGE_GUILD")) {
       message.delete().catch(() => {})
       return message.channel.send(`${message.author.toString()}, ${client.emoji.misc.xmark} **You aren't allowed to send multiple spoilers in this server**`).then(m => {
         setTimeout(() => {
@@ -135,14 +135,14 @@ module.exports = async (client, message) => {
     if(command.ignore) return;
 
     client.logger.cmd(`${message.author.username} used the command ${command.name}`)
-    let r = await msg.react('❌').catch(() => {})
+    let r = await msg.react('🗑️').catch(() => {})
     try {
-      let react = await r.message.awaitReactions((reaction, user) => reaction.emoji.name === "❌" && user.id === message.author.id, {
+      let react = await r.message.awaitReactions((reaction, user) => reaction.emoji.name === "🗑️" && user.id === message.author.id, {
         time: 10 * 60 * 1000,
         max: 1,
         errors: ['time']
       })
-      if (react.first().emoji.name === '❌') msg.delete()
+      if (react.first().emoji.name === '🗑️') msg.delete()
       client.logger.cmd(`${message.author.username} deleted the command usage ${command.name}`)
     } catch (e) {
       if(msg.embeds.length) {
